+++ NEWS.md
@@ -16,6 +16,12 @@
 
 * `ddbs_join()`: new function to perform spatial join operations.
 
+* `ddbs_length()`: adds a new column with the length of the geometries
+
+* `ddbs_area()`: adds a new column with the area of the geometries
+
+* `ddbs_is_valid()`: adds a new logical column asserting the validity of the geometries
+
 ## MINOR CHANGES
 
 * All functions now have a parameter `quiet` that allows users to suppress informational messages. Closed [#3](https://github.com/Cidree/duckspatial/issues/3)